Adding scripts to convert from dicoms to BIDS
These scripts require dcm2niix to work. The BidsConvert.sh script is the main script, while BidsConvertParameters.sh is where you define all the study specific parameters. There is a really hacky, non-elegant solution to add in a necessary statement to the .json files of the fieldmaps that still needs to be fixed.